@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import {DetailsList, IColumn, DetailsListLayoutMode, IDetailsHeaderProps, SelectionMode} from '@fluentui/react/lib/DetailsList';
+import {DetailsList, IColumn, DetailsListLayoutMode, IDetailsHeaderProps, SelectionMode, ConstrainMode} from '@fluentui/react/lib/DetailsList';
 import {mergeStyles, DefaultFontStyles } from '@fluentui/react/lib/Styling';
 import {initializeIcons} from '@fluentui/react/lib/Icons';
 import {ScrollablePane} from '@fluentui/react/lib/ScrollablePane';
@@ -84,7 +84,7 @@ export const ColorfulGrid = React.memo(function ColorfulGridApp({
         const isOptionSetRenderer : boolean = metadataAttributes?.has(column.original.name);      
         const columnDefaultIcon = displayIconType==="NAME" ? defaultIconNames.get(column.original.name)??defaultIcon : defaultIcon; 
         return {
-            ...getDefaultColumnSetup(column, dataset),
+            ...getDefaultColumnSetup(column, dataset),            
             onRender: isOptionSetRenderer===true  ? (item : any) => {      
               return <ColorfulCell 
                             item={item} 
@@ -112,6 +112,7 @@ export const ColorfulGrid = React.memo(function ColorfulGridApp({
                         selectionPreservedOnEmptyClick={true}
                         selectionMode={SelectionMode.multiple}     
                         layoutMode={DetailsListLayoutMode.justified}       
+                        constrainMode={ConstrainMode.unconstrained}
                         onItemInvoked={onItemInvoked}                        
                         ariaLabelForSelectionColumn="Toggle selection"
                         ariaLabelForSelectAllCheckbox="Toggle selection for all items"
