@@ -3,6 +3,7 @@ import * as React  from 'react';
 import { render } from 'react-dom';
 import { IGridColumn } from '../Generic/Hooks/useColumns';
 import { ISetupSchemaValue } from '../Utils/interfaces';
+import { ColorfulCellItem } from './ColorfulCellItem';
 
 export interface IColorfulCellProps {
     item: any;
@@ -14,37 +15,35 @@ export interface IColorfulCellProps {
 }
 
 export const ColorfulCell = function ColorfulCell({item, column, metadataOptions, displayTextType, displayIconType, defaultIcon} : IColorfulCellProps) : JSX.Element{    
-    const currentOptionSetValue=  item.raw.getValue(column.original.name) as number;    
-    let color = metadataOptions?.get(currentOptionSetValue?.toString() ?? "")?.color ?? "gray";  
-    if(color==="white"){
-        color = "gray"
+    
+    if(item.raw.getValue(column.original.name) ==null){
+        return <div></div>;
     }
-    const icon  = metadataOptions?.get(currentOptionSetValue?.toString() ?? "")?.icon ?? defaultIcon;  
-    const iconColor = displayTextType==="BOX" ? "white" : color;
-    const renderIcon = displayIconType!=="NONE" ? <Icon className="colorIcon" style={{color: iconColor , marginRight: "5px"}} iconName={icon} aria-hidden="true" /> : "";
-    const style = {
-        "BORDER" : {
-            borderWidth: "1px", 
-            borderStyle: "solid", 
-            borderColor: color, 
-            color: color,  
-            borderRadius: "5px"
-        },
-        "BOX" : {
-            backgroundColor: color==="white" ? "gray" : color, color: iconColor, borderRadius: "5px"
-        }, 
-        "SIMPLE" : {             
-        }, 
-        "NOTEXT": {       
-            cursor: "pointer"     
-        }
-    }[displayTextType];   
-    const content = item[column.original.name];
-    const renderText = displayTextType!=="NOTEXT" ? <span className="cell">{content}</span> : ""    
-    return(<div className="ColorfulCell" style={style} title={content}>            
-            {renderIcon}         
-            {renderText}
-        </div>);
+    if(column.original.dataType==="MultiSelectOptionSet"  || column.original.dataType==="MultiSelectPicklist"){
+        const currentValues = (item.raw.getValue(column.original.name) as string ?? "").split(",");        
+        const currentDisplayNames = (item.raw.getFormattedValue(column.original.name) as string ?? "").split(";");
+        return (<div className="ColorfulCell_MultiSelectOptionSet">
+            {currentValues.map((currentValue, index) => {
+               return (<ColorfulCellItem className='ColorfulCellItem' 
+               currentValue={parseInt(currentValue)} 
+               currentDisplayName={currentDisplayNames[index] ?? ""} 
+               defaultIcon={defaultIcon}
+               displayIconType={displayIconType}
+               displayTextType={displayTextType}
+               metadataOptions={metadataOptions}
+               />) 
+            })}
+        </div>)
+    }
+    const currentOptionSetValue=  item.raw.getValue(column.original.name) as number;    
+    return(<ColorfulCellItem className='ColorfulCell' 
+                currentValue={currentOptionSetValue} 
+                currentDisplayName={item[column.original.name]} 
+                defaultIcon={defaultIcon}
+                displayIconType={displayIconType}
+                displayTextType={displayTextType}
+                metadataOptions={metadataOptions}
+                />);
 
     
 };
